Replaced DataSources.xml by config.json
Also removed the xml reader stuff and introduced Config.java, a class
that can read the json config.

+++ src/main/java/de/tuberlin/dima/presslufthammer/transport/QueryHandler.java
@@ -38,12 +38,12 @@ public class QueryHandler {
     final Channel client;
     final QueryMessage queryMsg;
     final Query query;
-    private int numPartsExpected;
+    private long numPartsExpected;
     QueryStatus status;
     SchemaNode schema;
     List<InMemoryReadonlyTablet> parts;
 
-    public QueryHandler(int parts, QueryMessage queryMsg, SchemaNode schema,
+    public QueryHandler(long parts, QueryMessage queryMsg, SchemaNode schema,
             Channel client) {
         assert (queryMsg.getQueryId() > 0);
         this.parts = Lists.newLinkedList();
